feat(backend): address review

+++ python/beeai_framework/backend/types.py
@@ -111,7 +111,7 @@ ChatModelCache = BaseCache[list[ChatModelOutput]]
 
 
 class EmbeddingModelUsage(BaseModel):
-    tokens: float | None = None
+    tokens: int | None = None
 
 
 class EmbeddingModelInput(BaseModel):
